working

+++ ft_memset.c
@@ -9,11 +9,11 @@ void	*ft_memset(void *b, int c, size_t n)
 	mem = (unsigned char*)b;
 	if (n >= 8)
 	{
-		stamp_64 = 0xffU & c;
+		stamp_64 = 0xffU & (uint32_t)c;
 		stamp_64 = (stamp_64 << 0x8U) | stamp_64;
 		stamp_64 = (stamp_64 << 0x10U) | stamp_64;
 		stamp_64 = (stamp_64 << 0x20U) | stamp_64;
-		ptr_64 = (unsigned long long*)mem;
+		ptr_64 = (uint64_t*)mem;
 		while (n >= 8)
 		{
 			*ptr_64++ = stamp_64;
